backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
markdown==3.5.2
bleach==6.1.0
langchain==0.1.4
langgraph==0.0.20
httpx==0.26.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
alembic==1.13.1
python-dotenv==1.0.0
grpcio==1.76.0
protobuf==6.31.1
pgvector==0.2.4
python-multipart==0.0.6
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
boto3==1.34.14
groq==0.9.0
requests==2.31.0
pdfplumber==0.11.4
